feat: 멤버 생성 인자 수정 및 삭제 구현
