--- src/main/java/org/sopt/service/MemberService.java
@@ -1,15 +1,0 @@
-package org.sopt.service;
-
-import org.sopt.domain.Member;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MemberService {
-
-    Long join(String name);
-
-    Optional<Member> findOne(Long memberId);
-
-    List<Member> findAllMembers();
-}
